[VecOps] Base RVec implementation on LLVM's SmallVector
The new implementation introduces a small buffer optimization and
is not based on RAdoptAllocator anymore, but rather on LLVM's
SmallVector. On top of SmallVector, RVec adds (or better keeps)
the option to adopt an existing memory buffer.
Some std::vector methods are also added to SmallVector for backwards
compatibility of RVec.

This commit also removes the deprecated `RVec::AsVector`, which we
cannot implement with the same zero-copy semantics now that RVec
is not implemented in terms of a std::vector.