Record that the TClass is from GenerateInfoForPair.
The TClass generated by TStreamerInfo::GenerateInfoForPair is neither 'Interpreted' (cling
likely does not, yet, have any information about that pair) nor loaded (the user did not
request a dictionary for it) but need to have special treatment for those TClass.

It is not clear whether this should be a new state.  For now we introduce a new
data member (`fIsSyntheticPair`) and don't increase the size of TClass instance
by using a bit field.