[cling] CreateNestedNameSpecifierForScopeOf needs specialization context:
Instead of picking a "random" (the first) specialization when building the name
of a non-dependent member type of a template, pick the one that corresponds to
the "current" specialization. As CreateNestedNameSpecifierForScopeOf cannot know
what's "current", this context needs to be passed down / propagated.

Add a couple of asserts that the context is provided when needed, and is a
template specialization that matches the type at hand.

This fixes #7955.