[RF] Warning message on repeated named RooCmdArgs in RooCmdConfig
Emit warning message in RooCmdConfig when multiple named arguments of
the same type are encountered, for which no chaining behavior is
defined.

To make the warning message more clear, a map is implemeted to look up
the name of the function in the `RooFit` namespace that is used to
create the repeated RooCmdArg. Usually, the names of the argument and
the name of the function that created it are identical, but there are
exceptions that are hardcoded in this map.

Furthermore, a warning message is also emitted for the commands that
have a defined chaining behavior (`Import`, `Link`, and `Slice`), where
the alternatives that take a `std::map` with the multiple commands is
suggested. This is done to encourage RooFit code that can be easily
ported to Python if desired, because chainig command arguments with the
same name is not possible in Python RooFit (commands are passed as
keyword arguments, which can't be repeated).

This commit addresses Jira issue ROOT-2784.

Having such a warning message also helps users that try to chain more
commands than `RooAbsReal::plotOn` can digest. They will get the
solution in the warning message instead of having to ask in the forum,
which happend for example in:

https://root-forum.cern.ch/t/simultaneous-fit-with-8-pdfs-using-roosimultaneous-and-roodataset/45331