[RF] Make RooProdPdf::_pdfNSetList a std::vector and not RooLinkedList
Part of the RooProdPdf modernization with the intention of eventually
deprecating the RooLinkedList.

This change reduces the size of a RooProdPdf from 1600 to 1496 bytes.

A manual schema evolution rule to test the reading of older RooProdPdfs
is implemented in the `LinkDef.h`. The schma evolution is tested by
the stressRooFit tests, which are reading some RooProdPdfs from the
reference files.