[DF] Remove usages of Range operation in DistRDF
This commit changes the way in which the input dataset is split in logical ranges of entries. It also changes the subsequent creation of an RDataFrame instance inside a distributed task.
For a TTree based dataset, this happens through the creation of a TChain that represents the subset of the original dataset that contains the entries in the range.
For an empty source dataset, this happens by creating an RDataFrame with a number of entries

Before this commit:
For the TTree based case, the entries of the range are global with respect to the TChain created inside the distributed task.
`TChain.SetCacheEntryRange` is called with the start and end entries of the range to make sure we don't read more than needed
An RDataFrame is created with the chain as argument. The `Range` operation is used to select only the entries of the TChain in the current task that are inside the input entry range.

For the empty source case, an RDataFrame is created with the total number of entries decided in the user application.
The `Range` operation is used to limit the processing of the global RDataFrame to only the entries in the logical range

After this commit:
For the TTree based case, the input range will report global entries with respect to the TChain of the task. But it will also report local entries with respect to each single file of the TChain in the current task.
A global TEntryList is created and filled with one sub-TEntryList per file in the chain. This allows to do a preselection directly in the TChain of the entries that should be processed or not.
`SetCacheEntryRange` is still called to ensure we don't read more than needed (hence the need to still keep global entries in the range object)
The TChain (with the TEntryList) is passed as argument to an RDataFrame constructor.

For the empty source case, an RDataFrame is created with as many entries as there are in the current range. This invalidates usages of an index column like "rdfentry_" that assume all ranges refer to the same global number of entries.